last_commit

+++ django_herbs/apps/shop/views.py
@@ -1,10 +1,10 @@
 from django.shortcuts import render, redirect, HttpResponse
+from django.http import *
+
 from .models import *
 # Create your views here.
 from django.http import JsonResponse
-
 from django.middleware import csrf
-
 
 from django.contrib.auth.decorators import login_required
 
@@ -16,7 +16,7 @@ from django.core.exceptions import ObjectDoesNotExist
 @login_required
 def shop_product(request):
     '''Рендерит все продукты по id картинки
-    Нужно добавить красивую пагинацию
+    Нужно добавить красивую пагинацию js
     '''
     product_list = Shop_product.objects.all()
     image_list = Image_Product.objects.all()
@@ -57,10 +57,19 @@ def shop_product_id(request, product_id):
                                                                    'cartform': cartform, "image": image})
 
 
-
-# Функции класса Order
 @login_required
 def add_order(request):
+    '''
+    Обзор товаров в корзине.
+    Заполнение формы заказа.
+    Удаление сессионной корзины в случае подтвеждения заказа и перенаправление ко всем заказам и оплате.
+
+
+    :param request:
+    :return:
+    '''
+
+
     if request.method == "POST":
         order_form = OrderForm(data=request.POST)
         if order_form.is_valid():
@@ -69,8 +78,6 @@ def add_order(request):
             order.name = request.user
             order.status = Status.objects.get(id = 1)# СтатусP:Оформлен
             order.save()
-
-
             # Добавление товаров из сессии в заказ
             cart_set = Cart.objects.filter(session_key=request.session.session_key)
             for cart in cart_set:
@@ -82,36 +89,61 @@ def add_order(request):
                 )
             cart_set.delete()
 
+            order.save() # для обновлния поля общей суммы в методе save
+
             return redirect(show_orders)
+        else:
+            return HttpResponseBadRequest('<h1>Форма не прошла валидацию!Проверьте правильность введенных данных</h1><br>')
+
     else:
         cart_set = Cart.objects.filter(session_key= request.session.session_key)
         lock = cart_set.count()
         if lock == 0:
             return render(request,'shop/null_cart_error.html')
         else:
-
             shop_product = Shop_product.objects.all()
-
             order_form = OrderForm()
-
             return render(request, 'shop/order.html', context={'order_form': order_form, 'cart_set': cart_set,
                                                                'shop_product':shop_product})
 
-
 @login_required
 def show_orders(request):
-    '''Показывает все заказы пользователя и по POST запросу удаляет заказ'''
+    '''Показывает все заказы пользователя и по POST запросу:
+     1) удаляет заказ
+     2)перенаправляет к форме оплаты
+     '''
+
     if request.method == "POST":
-        order_id = request.POST["order_id"]
-        order = Order.objects.filter(id=order_id)
-        order.delete()
-        orders = Order.objects.filter(name=request.user)
-        return render(request, 'shop/checkout.html', context={'orders': orders})
+        if request.POST['order_id'][:3] == 'pay':
+            print(request.POST['order_id'][3:])
+            order_id = request.POST['order_id'][4:]
+
+            return redirect('/shop/pay/'+ str(order_id))
+
+        else:
+            order_id = request.POST["order_id"]
+            order = Order.objects.filter(id=order_id)
+            order.delete()
+
+            orders = Order.objects.filter(name=request.user)
+
+            return render(request, 'shop/checkout.html', context={'orders': orders} )
 
     else:
         orders = Order.objects.filter(name=request.user)
+
+
+
+
         return render(request, 'shop/checkout.html', context={'orders': orders})
 
-def pay_orders(request):
-    pass
+# Оплата заказа class PayOrder()
+def pay_orders(request, order_id):
+    if request.method == 'POST':
+        #
+        pass
+    # проверка оплаты -- смена статуса заказа -- перенаправление на страницу всех заказов
+
+    else:
+        return HttpResponseBadRequest('<h1>Форма---Оплата по карте---</h1><br>')
